+++ petalsonic-demo/src/gui/app.rs
@@ -869,7 +869,6 @@ impl eframe::App for SpatialAudioDemo {
                 petalsonic::PetalSonicEvent::RuntimeStateChanged(state) => {
                     log::info!("GUI: Audio runtime is now {state:?}");
                 }
-                _ => {}
             }
         }